Multi-cell post tasks to embrace nova-less UC
This repo is branchless, and needs to cover both deployment methods:
- With Nova on UC on Train,
- Without Nova on UC, as of Wallaby

Adapt the script to use inventory, if there is overcloud-deploy
directory prepared by tripleo, othwerwise generate inventory files as
it used to de earlier.

Also fail the script if any command fails. For example, ansible fails
if a cell already exists.

Signed-off-by: Bogdan Dobrelya <bdobreli@redhat.com>
Change-Id: I1cdf9d7ed2ab5e4fce39d0adb2db9f99e083ed8c