OSP17: Multi-cell post tasks to embrace nova-less UC
This repo is branchless, and needs to cover both deployment methods:
- With Nova on UC on Train,
- Without Nova on UC, as of Wallaby

Split the script to match osp16 and osp17 use cases.

In OSP17 cell1.internalapi /etc/hosts record is missing on cell1
controller, so switch fetching it from central controller, where it is
present.

Also adjust the osp17 version of the  script to fail, if any command
fails.
For example, ansible fails if a cell already exists.

Related: rhbz#2089512
Signed-off-by: Bogdan Dobrelya <bdobreli@redhat.com>
Change-Id: I1cdf9d7ed2ab5e4fce39d0adb2db9f99e083ed8c