Apply control properties to extra compute
In the event that there are no spare controllers when starting the
multi-cell scale up, reassign the extra compute to be a controller
instead. Currently when deploying with extra controllers IR fails during
the intropsection phase. When that is addressed will revert this
approach. Also update user login when using 17 since this has
transitioned from heat-admin to tripleo-admin.

Change-Id: Ie7471460e91be9714886b129828f2586a4c2dd25